requirements.txt
Flask==2.3.3
python-barcode==0.15.1
Pillow==10.0.1
Werkzeug==2.3.7